backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
motor==3.3.2
bcrypt==4.1.2
geopy==2.4.1
aiofiles==23.2.1
stripe==7.8.0
razorpay==1.4.2
aiohttp==3.9.1
